Move create and insert QueryBuilders to new package and add "category" attribute "realKnn" benchmarking tables
