+++ src/main/java/org/polypheny/simpleclient/scenario/vectorbench/queryBuilder/creation/CreateMetadata.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder;
+package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder.creation;
 
 import java.util.Map;
 import kong.unirest.core.HttpRequest;
