+++ src/main/java/org/polypheny/simpleclient/scenario/vectorbench/queryBuilder/insertion/InsertIntFeature.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder;
+package org.polypheny.simpleclient.scenario.vectorbench.queryBuilder.insertion;
 
 import com.google.gson.JsonObject;
 import java.util.Arrays;
